SIMSBIOHUB-428: Major security endpoint refactor
