--- api/src/paths/administrative/security/index.ts
@@ -1,121 +1,0 @@
-import { RequestHandler } from 'express';
-import { Operation } from 'express-openapi';
-import { SYSTEM_ROLE } from '../../../constants/roles';
-import { getDBConnection } from '../../../database/db';
-import { authorizeRequestHandler } from '../../../request-handlers/security/authorization';
-import { SecurityService } from '../../../services/security-service';
-import { getLogger } from '../../../utils/logger';
-
-const defaultLog = getLogger('paths/administrative/security');
-
-export const GET: Operation = [
-  authorizeRequestHandler(() => {
-    return {
-      and: [
-        {
-          validSystemRoles: [SYSTEM_ROLE.SYSTEM_ADMIN],
-          discriminator: 'SystemRole'
-        }
-      ]
-    };
-  }),
-  getActiveSecurityRules()
-];
-
-GET.apiDoc = {
-  description: 'Get all active security rules.',
-  tags: ['security'],
-  security: [
-    {
-      Bearer: []
-    }
-  ],
-  responses: {
-    200: {
-      description: 'Security Rules.',
-      content: {
-        'application/json': {
-          schema: {
-            type: 'array',
-            items: {
-              type: 'object',
-              properties: {
-                security_rule_id: {
-                  type: 'number'
-                },
-                name: {
-                  type: 'string'
-                },
-                description: {
-                  type: 'string'
-                },
-                record_effective_date: {
-                  type: 'string'
-                },
-                record_end_date: {
-                  type: 'string',
-                  nullable: true
-                },
-                create_date: {
-                  type: 'string'
-                },
-                create_user: {
-                  type: 'number'
-                },
-                update_date: {
-                  type: 'string',
-                  nullable: true
-                },
-                update_user: {
-                  type: 'number',
-                  nullable: true
-                },
-                revision_count: {
-                  type: 'number'
-                }
-              }
-            }
-          }
-        }
-      }
-    },
-    400: {
-      $ref: '#/components/responses/400'
-    },
-    401: {
-      $ref: '#/components/responses/401'
-    },
-    403: {
-      $ref: '#/components/responses/403'
-    },
-    500: {
-      $ref: '#/components/responses/500'
-    },
-    default: {
-      $ref: '#/components/responses/default'
-    }
-  }
-};
-
-export function getActiveSecurityRules(): RequestHandler {
-  return async (req, res) => {
-    const connection = getDBConnection(req['keycloak_token']);
-    const service = new SecurityService(connection);
-
-    try {
-      await connection.open();
-
-      const data = await service.getActiveSecurityRules();
-
-      await connection.commit();
-
-      return res.status(200).json(data);
-    } catch (error) {
-      defaultLog.error({ label: 'getActiveSecurityRules', message: 'error', error });
-      await connection.rollback();
-      throw error;
-    } finally {
-      connection.release();
-    }
-  };
-}
