+++ api/src/paths/administrative/security/submission/{submissionId}.ts
@@ -1,12 +1,115 @@
 import { RequestHandler } from 'express';
 import { Operation } from 'express-openapi';
-import { SYSTEM_ROLE } from '../../../constants/roles';
-import { getDBConnection } from '../../../database/db';
-import { authorizeRequestHandler } from '../../../request-handlers/security/authorization';
-import { SecurityService } from '../../../services/security-service';
-import { getLogger } from '../../../utils/logger';
+import { SYSTEM_ROLE } from '../../../../constants/roles';
+import { getDBConnection } from '../../../../database/db';
+import { authorizeRequestHandler } from '../../../../request-handlers/security/authorization';
+import { SecurityService } from '../../../../services/security-service';
+import { getLogger } from '../../../../utils/logger';
 
-const defaultLog = getLogger('paths/administrative/security/apply');
+const defaultLog = getLogger('paths/administrative/security/submission/{submissionId}');
+
+export const GET: Operation = [
+  authorizeRequestHandler(() => {
+    return {
+      and: [
+        {
+          validSystemRoles: [SYSTEM_ROLE.SYSTEM_ADMIN],
+          discriminator: 'SystemRole'
+        }
+      ]
+    };
+  }),
+  getAllSecurityRulesForSubmissionFeatures()
+];
+
+GET.apiDoc = {
+  description: '',
+  tags: ['security'],
+  security: [
+    {
+      Bearer: []
+    }
+  ],
+  parameters: [
+    {
+      description: 'Submission ID',
+      in: 'path',
+      name: 'submissionId',
+      schema: {
+        type: 'integer',
+        minimum: 1
+      },
+      required: true
+    }
+  ],
+  responses: {
+    200: {
+      description: 'Set of all security rules applied to all features belonging to the given submission.',
+      content: {
+        'application/json': {
+          schema: {
+            type: 'array',
+            items: {
+              type: 'object',
+              properties: {
+                submission_feature_security_id: {
+                  type: 'number'
+                },
+                submission_feature_id: {
+                  type: 'number'
+                },
+                security_rule_id: {
+                  type: 'number'
+                }
+              }
+            }
+          }
+        }
+      }
+    },
+    400: {
+      $ref: '#/components/responses/400'
+    },
+    401: {
+      $ref: '#/components/responses/401'
+    },
+    403: {
+      $ref: '#/components/responses/403'
+    },
+    500: {
+      $ref: '#/components/responses/500'
+    },
+    default: {
+      $ref: '#/components/responses/default'
+    }
+  }
+};
+
+export function getAllSecurityRulesForSubmissionFeatures(): RequestHandler {
+  return async (req, res) => {
+    const connection = getDBConnection(req['keycloak_token']);
+    const securityService = new SecurityService(connection);
+
+
+    try {
+      await connection.open();
+
+      const submissionId = Number(req.params.submissionId);
+      const data = await securityService.getAllSecurityRulesForSubmission(submissionId);
+
+      await connection.commit();
+
+      return res.status(200).json(data);
+    } catch (error) {
+      defaultLog.error({ label: 'getAllSecurityRulesForSubmissionFeatures', message: 'error', error });
+      await connection.rollback();
+      throw error;
+    } finally {
+      connection.release();
+    }
+  };
+}
+
 
 export const PATCH: Operation = [
   authorizeRequestHandler(() => {
